Add Apache 2 license (#13)
### What
Add Apache 2 license.

### Why
The repo is one of our public repos and in the readme it indicates the purpose is for people to take it and build on it so this should probably have an open source license. The license was taken from the stellar/go repository with its year set to 2019 since that was the year of the first commit to the repo.